services/crawler/requirements.txt
httpx==0.28.1
beautifulsoup4==4.12.3
langchain-text-splitters==0.3.11
python-dotenv==1.1.1
structlog==25.4.0
tenacity==9.1.2
pyarrow==18.1.0
